sys: types: long -> int

+++ sys/inc/sys/types.h
@@ -13,7 +13,7 @@
 /* Unsigned types */
 typedef unsigned char       __uint8_t;
 typedef unsigned short      __uint16_t;
-typedef unsigned long       __uint32_t;
+typedef unsigned int        __uint32_t;
 typedef unsigned long long  __uint64_t;
 typedef __uint64_t          __size_t;
 typedef __uint64_t          __uintptr_t;
@@ -21,7 +21,7 @@ typedef __uint64_t          __uintptr_t;
 /* Signed types */
 typedef int         __int8_t;
 typedef short       __int16_t;
-typedef long        __int32_t;
+typedef int         __int32_t;
 typedef long long   __int64_t;
 typedef __int64_t   __ssize_t;
 typedef __ssize_t   __off_t;
